add `fillDescriptions` method to MTDUncalibratedRecHitProducer
