+++ RecoLocalFastTime/FTLCommonAlgos/src/MTDUncalibratedRecHitAlgoBase.cc
@@ -1,4 +1,4 @@
 #include "RecoLocalFastTime/FTLCommonAlgos/interface/MTDUncalibratedRecHitAlgoBase.h"
-
-EDM_REGISTER_PLUGINFACTORY(BTLUncalibratedRecHitAlgoFactory, "BTLUncalibratedRecHitAlgoFactory");
-EDM_REGISTER_PLUGINFACTORY(ETLUncalibratedRecHitAlgoFactory, "ETLUncalibratedRecHitAlgoFactory");
+#include "FWCore/ParameterSet/interface/ValidatedPluginFactoryMacros.h"
+EDM_REGISTER_VALIDATED_PLUGINFACTORY(BTLUncalibratedRecHitAlgoFactory, "BTLUncalibratedRecHitAlgoFactory");
+EDM_REGISTER_VALIDATED_PLUGINFACTORY(ETLUncalibratedRecHitAlgoFactory, "ETLUncalibratedRecHitAlgoFactory");
